requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
tensorflow==2.15.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
pillow==10.0.0
python-multipart==0.0.6
pydantic==2.4.2